[esaggs][inspector]: Refactor to prep for esaggs move to server. (#83199)
