+++ docs/development/plugins/embeddable/public/kibana-plugin-plugins-embeddable-public.adapters.data.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-embeddable-public](./kibana-plugin-plugins-embeddable-public.md) &gt; [Adapters](./kibana-plugin-plugins-embeddable-public.adapters.md) &gt; [data](./kibana-plugin-plugins-embeddable-public.adapters.data.md)
+
+## Adapters.data property
+
+<b>Signature:</b>
+
+```typescript
+data?: DataAdapter;
+```
