+++ docs/development/plugins/embeddable/public/kibana-plugin-plugins-embeddable-public.adapters.md
@@ -11,3 +11,11 @@ The interface that the adapters used to open an inspector have to fullfill.
 ```typescript
 export interface Adapters 
 ```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [data](./kibana-plugin-plugins-embeddable-public.adapters.data.md) | <code>DataAdapter</code> |  |
+|  [requests](./kibana-plugin-plugins-embeddable-public.adapters.requests.md) | <code>RequestAdapter</code> |  |
+
