+++ src/plugins/inspector/common/adapters/index.ts
@@ -17,12 +17,6 @@
  * under the License.
  */
 
-export { Adapters } from './types';
-export { DataAdapter, FormattedData } from './data';
-export {
-  RequestAdapter,
-  RequestStatistic,
-  RequestStatistics,
-  RequestStatus,
-  RequestResponder,
-} from './request';
+export * from './data';
+export * from './request';
+export * from './types';
