+++ src/plugins/inspector/common/adapters/types.ts
@@ -17,9 +17,14 @@
  * under the License.
  */
 
+import type { DataAdapter } from './data';
+import type { RequestAdapter } from './request';
+
 /**
  * The interface that the adapters used to open an inspector have to fullfill.
  */
 export interface Adapters {
+  data?: DataAdapter;
+  requests?: RequestAdapter;
   [key: string]: any;
 }
