+++ src/plugins/inspector/public/views/data/components/data_view.tsx
@@ -35,7 +35,7 @@ import { Adapters } from '../../../../common';
 import {
   TabularLoaderOptions,
   TabularData,
-  TabularCallback,
+  TabularHolder,
 } from '../../../../common/adapters/data/types';
 import { IUiSettingsClient } from '../../../../../../core/public';
 import { withKibana, KibanaReactContextValue } from '../../../../../kibana_react/public';
@@ -44,7 +44,7 @@ interface DataViewComponentState {
   tabularData: TabularData | null;
   tabularOptions: TabularLoaderOptions;
   adapters: Adapters;
-  tabularPromise: TabularCallback | null;
+  tabularPromise: Promise<TabularHolder> | null;
 }
 
 interface DataViewComponentProps extends InspectorViewProps {
@@ -73,7 +73,7 @@ class DataViewComponent extends Component<DataViewComponentProps, DataViewCompon
       adapters: nextProps.adapters,
       tabularData: null,
       tabularOptions: {},
-      tabularPromise: nextProps.adapters.data.getTabular(),
+      tabularPromise: nextProps.adapters.data!.getTabular(),
     };
   }
 
@@ -82,7 +82,7 @@ class DataViewComponent extends Component<DataViewComponentProps, DataViewCompon
       this.setState({
         tabularData: null,
         tabularOptions: {},
-        tabularPromise: this.props.adapters.data.getTabular(),
+        tabularPromise: this.props.adapters.data!.getTabular(),
       });
     }
   };
@@ -91,7 +91,7 @@ class DataViewComponent extends Component<DataViewComponentProps, DataViewCompon
     const { tabularPromise } = this.state;
 
     if (tabularPromise) {
-      const tabularData: TabularData = await tabularPromise;
+      const tabularData: TabularHolder = await tabularPromise;
 
       if (this._isMounted) {
         this.setState({
@@ -105,13 +105,13 @@ class DataViewComponent extends Component<DataViewComponentProps, DataViewCompon
 
   componentDidMount() {
     this._isMounted = true;
-    this.props.adapters.data.on('change', this.onUpdateData);
+    this.props.adapters.data!.on('change', this.onUpdateData);
     this.finishLoadingData();
   }
 
   componentWillUnmount() {
     this._isMounted = false;
-    this.props.adapters.data.removeListener('change', this.onUpdateData);
+    this.props.adapters.data!.removeListener('change', this.onUpdateData);
   }
 
   componentDidUpdate() {
